+++ execution-core/tests/serialization.rs
@@ -7,14 +7,15 @@
 use dusk_bls12_381::BlsScalar;
 use dusk_bytes::{Error, Serializable};
 use dusk_jubjub::JubJubScalar;
-use execution_core::transfer::{ContractCall, Fee, Payload, Transaction};
+use execution_core::transfer::{
+    CallOrDeploy, ContractCall, ContractDeploy, Fee, Payload, Transaction,
+};
 use execution_core::{Note, PublicKey, SecretKey, TxSkeleton};
 use ff::Field;
 use rand::rngs::StdRng;
 use rand::SeedableRng;
 
-#[test]
-fn transaction_from_to_bytes() -> Result<(), Error> {
+fn build_skeleton_fee_deposit() -> (TxSkeleton, Fee, u64) {
     let mut rng = StdRng::seed_from_u64(42);
 
     // set the general parameters
@@ -61,6 +62,12 @@ fn transaction_from_to_bytes() -> Result<(), Error> {
 
     // build the fee
     let fee = Fee::new(&mut rng, &sender_pk, gas_limit, gas_price);
+    (tx_skeleton, fee, deposit)
+}
+
+#[test]
+fn transaction_serialization_call() -> Result<(), Error> {
+    let (tx_skeleton, fee, deposit) = build_skeleton_fee_deposit();
 
     // build the contract-call
     let contract = [42; 32];
@@ -74,7 +81,41 @@ fn transaction_from_to_bytes() -> Result<(), Error> {
     let payload = Payload {
         tx_skeleton,
         fee,
-        contract_call: Some(call),
+        call_or_deploy: Some(CallOrDeploy::Call(call)),
+    };
+
+    // set a random proof
+    let proof = [42; 42].to_vec();
+
+    let transaction = Transaction::new(payload, proof);
+
+    let transaction_bytes = transaction.to_var_bytes();
+    let deserialized = Transaction::from_slice(&transaction_bytes)?;
+    assert_eq!(transaction, deserialized);
+    Ok(())
+}
+
+#[test]
+fn transaction_serialization_deploy() -> Result<(), Error> {
+    let (tx_skeleton, fee, _) = build_skeleton_fee_deposit();
+
+    // build the contract-deploy
+    let contract = [42; 32];
+    let bytecode = vec![1, 2, 3, 4, 5];
+    let owner = [1; 32];
+    let constructor_args = vec![5];
+    let deploy = ContractDeploy {
+        contract_id: Some(contract),
+        bytecode,
+        owner: owner.to_vec(),
+        constructor_args: Some(constructor_args),
+    };
+
+    // build the payload
+    let payload = Payload {
+        tx_skeleton,
+        fee,
+        call_or_deploy: Some(CallOrDeploy::Deploy(deploy)),
     };
 
     // set a random proof